// frontend/src/types/CompanyDetailsType.ts
export interface CompanyDetails {
  companyName: string;
  gstin: string;
  pan: string;
  email: string;
  phone: string;
  website: string;
  address: string;
  city: string;
  state: string;
  postalCode: string;
  bankName: string;
  accountHolderName: string;
  accountNumber: string;
  accountType: string;
  ifscCode: string;
  bankBranch: string;
}